refactor(tui): organize modules by feature
