+++ src/borgboi/tui/__init__.py
@@ -1,9 +1,8 @@
 """BorgBoi Textual TUI."""
 
 from borgboi.tui.app import BorgBoiApp
-from borgboi.tui.config_panel import ConfigPanel
-from borgboi.tui.config_screen import ConfigScreen
-from borgboi.tui.daily_backup_screen import DailyBackupScreen
-from borgboi.tui.excludes_screen import DefaultExcludesScreen
+from borgboi.tui.features.daily_backup import DailyBackupScreen
+from borgboi.tui.screens import ConfigScreen, DefaultExcludesScreen
+from borgboi.tui.widgets import ConfigPanel
 
 __all__ = ["BorgBoiApp", "ConfigPanel", "ConfigScreen", "DailyBackupScreen", "DefaultExcludesScreen"]
