+++ src/borgboi/tui/features/__init__.py
@@ -1,0 +1,1 @@
+"""Feature packages for BorgBoi's Textual UI."""
