+++ src/borgboi/tui/features/archive_compare/__init__.py
@@ -1,0 +1,23 @@
+"""Archive comparison TUI feature."""
+
+from borgboi.tui.features.archive_compare.content_diff_modal import ContentDiffScreen
+from borgboi.tui.features.archive_compare.screen import (
+    ArchiveCompareScreen,
+    CompareDirectoryTree,
+    ComparePathState,
+    build_compare_path_states,
+    build_compare_tree_highlights,
+    build_compare_tree_modified_paths,
+    build_compare_tree_parent_indicators,
+)
+
+__all__ = [
+    "ArchiveCompareScreen",
+    "CompareDirectoryTree",
+    "ComparePathState",
+    "ContentDiffScreen",
+    "build_compare_path_states",
+    "build_compare_tree_highlights",
+    "build_compare_tree_modified_paths",
+    "build_compare_tree_parent_indicators",
+]
