+++ src/borgboi/tui/features/repo_detail/__init__.py
@@ -1,0 +1,16 @@
+"""Repository detail TUI feature."""
+
+from borgboi.tui.features.repo_detail.config_screen import RepoConfigResult, RepoConfigScreen
+from borgboi.tui.features.repo_detail.info_screen import RepoInfoScreen, load_repo_excludes_state
+from borgboi.tui.features.repo_detail.quota import effective_quota_display
+from borgboi.tui.features.repo_detail.workspace import RepoWorkspaceState, load_repo_workspace_state
+
+__all__ = [
+    "RepoConfigResult",
+    "RepoConfigScreen",
+    "RepoInfoScreen",
+    "RepoWorkspaceState",
+    "effective_quota_display",
+    "load_repo_excludes_state",
+    "load_repo_workspace_state",
+]
