+++ src/borgboi/tui/features/repo_detail/workspace.py
