+++ src/borgboi/tui/screens/__init__.py
@@ -1,0 +1,6 @@
+"""Standalone TUI screens."""
+
+from borgboi.tui.screens.config import ConfigScreen
+from borgboi.tui.screens.excludes import DefaultExcludesScreen
+
+__all__ = ["ConfigScreen", "DefaultExcludesScreen"]
