+++ src/borgboi/tui/support/__init__.py
@@ -1,0 +1,5 @@
+"""Cross-cutting TUI support helpers."""
+
+from borgboi.tui.support.telemetry import capture_span
+
+__all__ = ["capture_span"]
