+++ src/borgboi/tui/widgets/__init__.py
@@ -1,0 +1,5 @@
+"""Reusable TUI widgets."""
+
+from borgboi.tui.widgets.config_panel import ConfigPanel
+
+__all__ = ["ConfigPanel"]
